requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0

# AI Framework
crewai>=0.11.2
langchain>=0.1.0
openai>=1.0.0

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0

# Trading
ccxt>=4.0.0

# Utils
aiohttp>=3.9.0
requests>=2.31.0